refactor(mdx/basic): merge contentful-image with image component
