--- mdx/contentful/CHANGELOG.md
@@ -1,127 +1,0 @@
-# Change Log
-
-All notable changes to this project will be documented in this file.
-See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
-
-## [0.1.7](https://github.com/axe312ger/gatsby-mdx-suite/compare/@gatsby-mdx-suite/mdx-contentful@0.1.6...@gatsby-mdx-suite/mdx-contentful@0.1.7) (2019-12-12)
-
-**Note:** Version bump only for package @gatsby-mdx-suite/mdx-contentful
-
-
-
-
-
-## [0.1.6](https://github.com/axe312ger/gatsby-mdx-suite/compare/@gatsby-mdx-suite/mdx-contentful@0.1.5...@gatsby-mdx-suite/mdx-contentful@0.1.6) (2019-12-12)
-
-**Note:** Version bump only for package @gatsby-mdx-suite/mdx-contentful
-
-
-
-
-
-## [0.1.5](https://github.com/axe312ger/gatsby-mdx-suite/compare/@gatsby-mdx-suite/mdx-contentful@0.1.4...@gatsby-mdx-suite/mdx-contentful@0.1.5) (2019-12-10)
-
-**Note:** Version bump only for package @gatsby-mdx-suite/mdx-contentful
-
-
-
-
-
-## [0.1.4](https://github.com/axe312ger/gatsby-mdx-suite/compare/@gatsby-mdx-suite/mdx-contentful@0.1.3...@gatsby-mdx-suite/mdx-contentful@0.1.4) (2019-12-10)
-
-**Note:** Version bump only for package @gatsby-mdx-suite/mdx-contentful
-
-
-
-
-
-## [0.1.3](https://github.com/axe312ger/gatsby-mdx-suite/compare/@gatsby-mdx-suite/mdx-contentful@0.1.2...@gatsby-mdx-suite/mdx-contentful@0.1.3) (2019-12-10)
-
-**Note:** Version bump only for package @gatsby-mdx-suite/mdx-contentful
-
-
-
-
-
-## [0.1.2](https://github.com/axe312ger/gatsby-mdx-suite/compare/@gatsby-mdx-suite/mdx-contentful@0.1.1...@gatsby-mdx-suite/mdx-contentful@0.1.2) (2019-10-29)
-
-
-### Bug Fixes
-
-* **contentful:** ensure img and svg width is based on container ([a433fb1](https://github.com/axe312ger/gatsby-mdx-suite/commit/a433fb1f5beb33f9dc523b9c09ce748656961dee))
-
-
-
-
-
-## [0.1.1](https://github.com/axe312ger/gatsby-mdx-suite/compare/@gatsby-mdx-suite/mdx-contentful@0.1.0...@gatsby-mdx-suite/mdx-contentful@0.1.1) (2019-10-29)
-
-
-### Bug Fixes
-
-* **contentful:** return null when image data is not available ([b69ec9e](https://github.com/axe312ger/gatsby-mdx-suite/commit/b69ec9e7e881cfa21915469da6e93aa78c9de5ad))
-
-
-
-
-
-# [0.1.0](https://github.com/axe312ger/gatsby-mdx-suite/compare/@gatsby-mdx-suite/mdx-contentful@0.0.8...@gatsby-mdx-suite/mdx-contentful@0.1.0) (2019-10-29)
-
-
-### Features
-
-* **contentful:** allow contentful image to load data from different context key ([1858ab2](https://github.com/axe312ger/gatsby-mdx-suite/commit/1858ab25a0df35d523587bbd61ee13d9cbafd02c))
-
-
-
-
-
-## [0.0.8](https://github.com/axe312ger/gatsby-mdx-suite/compare/@gatsby-mdx-suite/mdx-contentful@0.0.7...@gatsby-mdx-suite/mdx-contentful@0.0.8) (2019-10-29)
-
-**Note:** Version bump only for package @gatsby-mdx-suite/mdx-contentful
-
-
-
-
-
-## [0.0.7](https://github.com/axe312ger/gatsby-mdx-suite/compare/@gatsby-mdx-suite/mdx-contentful@0.0.6...@gatsby-mdx-suite/mdx-contentful@0.0.7) (2019-10-29)
-
-**Note:** Version bump only for package @gatsby-mdx-suite/mdx-contentful
-
-
-
-
-
-## [0.0.6](https://github.com/axe312ger/gatsby-mdx-suite/compare/@gatsby-mdx-suite/mdx-contentful@0.0.5...@gatsby-mdx-suite/mdx-contentful@0.0.6) (2019-10-21)
-
-
-### Bug Fixes
-
-* **docs:** move docs into mdx modules ([029723f](https://github.com/axe312ger/gatsby-mdx-suite/commit/029723fbe0a1630b91ac480e419e1479459ad472))
-
-
-
-
-
-## [0.0.5](https://github.com/axe312ger/gatsby-mdx-suite/compare/@gatsby-mdx-suite/mdx-contentful@0.0.4...@gatsby-mdx-suite/mdx-contentful@0.0.5) (2019-10-21)
-
-**Note:** Version bump only for package @gatsby-mdx-suite/mdx-contentful
-
-
-
-
-
-## [0.0.4](https://github.com/axe312ger/gatsby-mdx-suite/compare/@gatsby-mdx-suite/mdx-contentful@0.0.3...@gatsby-mdx-suite/mdx-contentful@0.0.4) (2019-10-21)
-
-**Note:** Version bump only for package @gatsby-mdx-suite/mdx-contentful
-
-
-
-
-
-## [0.0.3](https://github.com/axe312ger/gatsby-mdx-suite/compare/@gatsby-mdx-suite/mdx-contentful@0.0.2...@gatsby-mdx-suite/mdx-contentful@0.0.3) (2019-10-21)
-
-
-### Bug Fixes
-
-* **mdx-basic:** move gatsby image wrapper to basic package and fix depencencies ([cf79010](https://github.com/axe312ger/gatsby-mdx-suite/commit/cf790102c84d4ddbeb4180eec85504030b7b5ecd))
