--- mdx/contentful/index.js
@@ -1,1 +1,0 @@
-// noop
