--- mdx/contentful/package.json
@@ -1,26 +1,0 @@
-{
-  "name": "@gatsby-mdx-suite/mdx-contentful",
-  "version": "0.1.7",
-  "description": "",
-  "author": "Benedikt Rötsch <opensource@axe312.dev>",
-  "license": "MIT",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "keywords": [
-    "gatsby-mdx-suite",
-    "gatsby",
-    "mdx"
-  ],
-  "dependencies": {
-    "@emotion/styled": "^10.0.23",
-    "@gatsby-mdx-suite/contexts": "^0.0.4",
-    "@gatsby-mdx-suite/mdx-basic": "^0.0.13",
-    "prop-types": "^15.7.2",
-    "react": "^16.12.0"
-  },
-  "publishConfig": {
-    "access": "public"
-  }
-}
